+++ app/schemas/discogs.schemas.ts
@@ -1,0 +1,14 @@
+import z from "zod";
+
+const DiscogsResponseSchema = z.object({
+  results: z.array(
+    z.object({
+      id: z.number(),
+      cover_image: z.string(),
+      title: z.string(),
+      year: z.string().optional(),
+    })
+  ),
+});
+
+export { DiscogsResponseSchema };
