+++ components/Result.tsx
@@ -1,11 +1,11 @@
 import Image from "next/image";
 import Icon from "./Icon";
-import { Album } from "@/app/utils/discogs.schemas";
+import { Album } from "@/app/schemas/collections.schemas";
 
 export default function Result({ album }: { album: Album }) {
   return (
     <div className="flex p-24 border-bottom-solid border-b-2 border-stone-900">
-      {album.cover_image.endsWith(".gif") ? (
+      {album.coverImage.endsWith(".gif") ? (
         <div className="w-[200px] h-[200px] p-12 rounded-sm bg-gray-800">
           <Icon
             className="w-full h-full text-white "
@@ -16,7 +16,7 @@ export default function Result({ album }: { album: Album }) {
       ) : (
         <Image
           className="rounded-sm"
-          src={album.cover_image}
+          src={album.coverImage}
           alt={`${album.albumTitle} cover image`}
           width={200}
           height={200}
